Implement colorspace generation in engine, generate colorspace from nclx.
The engine will first ask the registery if it has matching colorspaces,
else, it'll generate a colorspace and add it to the registery.

The unittests can't be linked properly to the things I want to test,
so I just manually test the code that I need, which has been useful
for figuring out the appropriate way to generate certain transfer
functions.

+++ libs/pigment/KoColorProfileStorage.cpp
@@ -135,7 +135,7 @@ QList<const KoColorProfile *> KoColorProfileStorage::profilesFor(const KoColorSp
     return profiles;
 }
 
-QList<const KoColorProfile *> KoColorProfileStorage::profilesFor(QVector<float> colorants, int colorantType, int transferType, double error)
+QList<const KoColorProfile *> KoColorProfileStorage::profilesFor(QVector<double> colorants, int colorantType, int transferType, double error)
 {
     QList<const KoColorProfile *>  profiles;
